+++ src/steps/utils.ts
@@ -21,20 +21,24 @@ export class VulnerabilityFindingKeysCollector {
     this.mapping = new Map();
   }
 
+  public allQids(): number[] {
+    return Array.from(this.mapping.keys());
+  }
+
   /**
    * Adds a Finding._key to set of vulnerability findings.
    *
    * @param qid vulnerability QID
    * @param findingKey Finding._key related to vulnerability
    */
-  public addVulnerabilityFinding(qid: number, findingKey: string): void {
+  public addVulnerabilityFindingKey(qid: number, findingKey: string): void {
     if (!qid)
       throw new IntegrationError({
         code: 'UNDEFINED_VULNERABILITY_QID',
         message: 'undefined QID provided for vulnerability',
       });
 
-    let keys = this.mapping[qid];
+    let keys = this.mapping.get(qid);
     if (!keys) {
       keys = new Set();
       this.mapping.set(qid, keys);
@@ -42,12 +46,32 @@ export class VulnerabilityFindingKeysCollector {
     keys.add(findingKey);
   }
 
+  public getVulnerabiltyFindingKeys(qid: number): string[] {
+    return Array.from(this.mapping.get(qid) || []);
+  }
+
   /**
    * Serializes collected values into form that can be stored between steps and
    * used to re-create the Map.
    */
-  public toVulnerabilityFindingKeys(): SerializedVulnerabilityFindingKeys {
+  public serialize(): SerializedVulnerabilityFindingKeys {
     return Array.from(this.mapping.entries());
+  }
+
+  /**
+   * Adds values previously serialized.
+   */
+  public loadSerialized(serialized: SerializedVulnerabilityFindingKeys) {
+    for (const [qid, findingKeys] of serialized) {
+      let existingFindingKeys = this.mapping.get(qid);
+      if (!existingFindingKeys) {
+        existingFindingKeys = new Set();
+        this.mapping.set(qid, existingFindingKeys);
+      }
+      for (const findingKey of findingKeys) {
+        existingFindingKeys.add(findingKey);
+      }
+    }
   }
 }